refactor(web): move share final comments code inline with share LPA
Currently there are separate implementations for LPA statements/comments
and final comments. These can be brought under a single implementation.
