--- appeals/web/src/server/appeals/appeal-details/progress-case/from-final-comments/from-final-comments.router.js
@@ -1,13 +1,0 @@
-import { Router as createRouter } from 'express';
-import { asyncHandler } from '@pins/express';
-
-import * as controller from './from-final-comments.controller.js';
-
-const router = createRouter({ mergeParams: true });
-
-router
-	.route('/')
-	.get(asyncHandler(controller.renderProgressFromFinalComments))
-	.post(asyncHandler(controller.postProgressFromFinalComments));
-
-export default router;
