--- appeals/web/src/server/appeals/appeal-details/progress-case/progress-case.router.js
@@ -1,9 +1,0 @@
-import { Router as createRouter } from 'express';
-import { validateAppeal } from '#appeals/appeal-details/appeal-details.middleware.js';
-import fromFinalCommentsRouter from './from-final-comments/from-final-comments.router.js';
-
-const router = createRouter({ mergeParams: true });
-
-router.use('/from-final-comments', validateAppeal, fromFinalCommentsRouter);
-
-export default router;
